LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/parsing/position-try-parsing-expected.txt


FAIL e.style['position-try'] = "flip-inline" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['position-try'] = "most-height none" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['position-try'] = "--bar, --baz" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['position-try'] = "most-inline-size --baz, flip-inline" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['position-try'] = "most-block-size flip-inline flip-block, --bar, --baz" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['position-try'] = "normal none" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['position-try'] = "most-width none" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['position-try'] = "normal --foo" should set the property value assert_not_equals: property should be set got disallowed value ""
PASS e.style['position-try'] = "normal --foo, most-width --bar" should not set the property value
PASS e.style['position-try'] = "none normal" should not set the property value
PASS e.style['position-try'] = "flip-block most-height" should not set the property value
PASS e.style['position-try'] = "most-height, flip-start" should not set the property value
FAIL e.style['position-try'] = "flip-inline" should set position-try-fallbacks assert_equals: position-try-fallbacks should be canonical expected "flip-inline" but got ""
FAIL e.style['position-try'] = "flip-inline" should set position-try-order assert_equals: position-try-order should be canonical expected "normal" but got ""
FAIL e.style['position-try'] = "flip-inline" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['position-try'] = "most-width none" should set position-try-fallbacks assert_equals: position-try-fallbacks should be canonical expected "none" but got ""
FAIL e.style['position-try'] = "most-width none" should set position-try-order assert_equals: position-try-order should be canonical expected "most-width" but got ""
FAIL e.style['position-try'] = "most-width none" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['position-try'] = "--foo, --bar" should set position-try-fallbacks assert_equals: position-try-fallbacks should be canonical expected "--foo, --bar" but got ""
FAIL e.style['position-try'] = "--foo, --bar" should set position-try-order assert_equals: position-try-order should be canonical expected "normal" but got ""
FAIL e.style['position-try'] = "--foo, --bar" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['position-try'] = "most-inline-size --foo, flip-inline" should set position-try-fallbacks assert_equals: position-try-fallbacks should be canonical expected "--foo, flip-inline" but got ""
FAIL e.style['position-try'] = "most-inline-size --foo, flip-inline" should set position-try-order assert_equals: position-try-order should be canonical expected "most-inline-size" but got ""
FAIL e.style['position-try'] = "most-inline-size --foo, flip-inline" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['position-try'] = "most-inline-size flip-inline flip-block, --foo, --bar" should set position-try-fallbacks assert_equals: position-try-fallbacks should be canonical expected "flip-inline flip-block, --foo, --bar" but got ""
FAIL e.style['position-try'] = "most-inline-size flip-inline flip-block, --foo, --bar" should set position-try-order assert_equals: position-try-order should be canonical expected "most-inline-size" but got ""
FAIL e.style['position-try'] = "most-inline-size flip-inline flip-block, --foo, --bar" should not set unrelated longhands assert_true: expected true got false
